+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.0.3
 
-- Change name to Music Player
+- Add possibility to seek on iOS.
+- Properly send the `.onCompleted` event on iOS.
+- Change name to Music Player.
+
 
 ## 0.0.2
 
@@ -9,4 +12,4 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
-- The initial release with a working version of the iOS player
+- The initial release with a working version of the iOS player.